libsks: update PKCS#11 header files
This change provides the PKCS#11 header file delivery as a single
pkcs11.h header file under a BSD type license for the OP-TEE project.
This header defines all API function prototypes and not all the Cryptoki
API:
- all token/session management structure and IDs are defined.
- all object attribute IDs define by PKCS#11 2.40-e01
- Some mechanisms IDs and their relates IDs/structure are not defined,
  restricting to AES support.

It is expected that future changes will add required definitions as
more features will be supported.

This change updates the libsks implementation according to this header
files replacement.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>