C_GetAttributeValue and C_GenerateKeyPair attribute guessing
Signed-off-by: Gabor Szekely <szvgabor@gmail.com>

- Added first implementation of C_GetAttributeValue, for this I
implemented a deserialize_ck_attribute function, however some
cases are not handled there. The return values are also not
always correct. I will deal with these issues later.

- Added key type guessing to C_GenerateKeyPair, since the PKCS11
standard states that in cases when the key type is deducable from
the key generation mechanism, it is not required to specify it.
For now, only one type of key is guessed, but more can be added
later, and the same guessing should be done in C_GenerateKey