libsks: guess key type on obvious cases
Some tools as pkcs11-tool do not specify some object attributes
when their value is obvious according to the context, such as the
object key type when mechanism used is a specifc  key generation.

This change guesses key type for RSA key pair generation when the
key type is not defined in the object attributes.

Signed-off-by: Gabor Szekely <szvgabor@gmail.com>
[Extracted as single commit, minor fixes]
Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>